pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auditkit"
version = "0.1.0"
description = "Grounded answers, control mapping, and audit-ready exports from bounded source material."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.54,<2",
  "openai>=2,<3",
  "pydantic>=2,<3",
  "python-dotenv>=1,<2",
  "pypdf>=5,<6",
  "faiss-cpu>=1.13,<2",
  "numpy>=2,<3",
  "httpx>=0.28,<1",
  "rich>=14,<15",
  "pytest>=8,<9",
  "ruff>=0.11,<1",
  "json-repair>=0.57,<1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
